move tests to general unit test package and refactor some logic extending from GraphQLTestBase
