+++ vertx-web-graphql/src/test/java/io/vertx/ext/web/handler/graphql/Musketeer.java
@@ -1,0 +1,7 @@
+package io.vertx.ext.web.handler.graphql;
+
+public enum Musketeer {
+  ATHOS,
+  PORTHOS,
+  ARAMIS;
+}
