--- vertx-web-graphql/src/test/java/io/vertx/ext/web/handler/graphql/it/Musketeer.java
@@ -1,7 +1,0 @@
-package io.vertx.ext.web.handler.graphql.it;
-
-enum Musketeer {
-  ATHOS,
-  PORTHOS,
-  ARAMIS;
-}
